Curl_range: commonize FTP and FILE range handling
